fold breakpoint manager into debug session
